Moves S3Utilities methods into sharedutils so they can be used in rda pipeline.
